requirements.txt
accelerate==1.1.1
annotated-types==0.7.0
bitsandbytes==0.45.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decord==0.6.0
deepspeed==0.15.1
diffusers @ git+https://github.com/huggingface/diffusers@de6a88c2d7659c616b44c0856677335110b8ff2e
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
hjson==3.1.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.6.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.4
nvtx==0.2.10
opencv-python==4.10.0.84
optimum==1.24.0
optimum-quanto==0.2.6
packaging==24.2
para_attn @ git+https://github.com/Howe2018/ParaAttention.git@414d8b8f503232bada4bcb6b6d6534c31a550e6a
peft==0.15.0
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sageattention==1.0.6
scipy==1.15.2
sentry-sdk==2.23.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
xformers==0.0.29.post1
xxhash==3.5.0
zipp==3.21.0
